backend/backendApp/current_requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
joblib==1.3.2
numpy==1.24.4
pip==24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
scikit-learn==1.5.2
scipy==1.14.1
setuptools==73.0.1
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
wheel==0.44.0
